requirements.txt
hazm==0.7.0
wordcloud==1.8.1

python-bidi==0.4.2
arabic-reshaper==2.1.3
loguru==0.5.3